+++ resources/views/billing-profile/delete.blade.php
@@ -2,14 +2,14 @@
 
 @section('content')
     <div class="container">
-        <h1>Elimina <em>{{ $company->business_name }}</em></h1>
-        <form method="POST" action="/companies/{{ $company->id }}">
+        <h1>Elimina <em>{{ $billingProfile->name }}</em></h1>
+        <form method="POST" action="{{ route('billing-profiles.destroy', ['billing_profile' => $billingProfile->id]) }}">
             @csrf
             @method('DELETE')
 
-            <p>Sei sicuro di voler eliminare {{ $company->business_name }}? Questa azione è <em>irreversibile</em>.</p>
+            <p>Sei sicuro di voler eliminare {{ $billingProfile->name }}? Questa azione è <em>irreversibile</em>.</p>
             <button type="submit" class="btn btn-primary">Elimina</button>
-            <a href="{{ route('companies.index') }}" class="btn btn-link">Annulla</a>
+            <a href="{{ route('billing-profiles.index') }}" class="btn btn-link">Annulla</a>
         </form>
     </div>
 @endsection